removed lots of stuff from base.py
