+++ planzero/base.py
@@ -266,13 +266,6 @@ class State(object):
         self._t_now = t_start # always in some unit of time, not always the same unit
         self.sts = {}  # the sts objects built up by rolling simulation forward
 
-        # TODO: index by enums.IPCC_Sector
-        self.sectoral_emissions_contributors = {
-            catpath: {}
-            for catpath in sorted(ipcc_canada.catpaths)}
-
-        self.subsidy_requirements = set()
-
         self.projects = {}
         self.project_writes = {} # prj.identifier -> set of string names
         self.project_requires_current = {} # prj.identifier -> set of string names
@@ -423,14 +416,6 @@ class State(object):
     def t_now(self, t_next):
         assert t_next >= self._t_now
         self._t_now = t_next
-
-    def register_emission(self, category_path, ghg, sts_key):
-        if self.emissions_registration_closed:
-            raise RuntimeError()
-        assert ghg == GHG(ghg)
-        assert sts_key in self.sts
-        self.sectoral_emissions_contributors[category_path]\
-                .setdefault(ghg, []).append(sts_key)
 
     def register_emission_factor(self, pt, driver, sts_key, ipcc_sector, ghg):
         if self.emissions_registration_closed:
@@ -621,422 +606,6 @@ class State(object):
 
 Scenario = State
 
-
-surface_area_of_earth = 5.1e14 * u.m * u.m
-
-molar_mass_CH4 = 16.0 * u.g / u.mol
-molar_mass_CO2 = 44.0 * u.g / u.mol
-molar_mass_N2O = 44.01 * u.g / u.mol
-
-atmospheric_conc_per_mass_CO2 = (1 * u.ppm) / (7.8 * u.gigatonne_CO2)
-atmospheric_conc_per_mass_CH4 = (1 * u.ppm) / (7.8 * u.gigatonne_CH4) * molar_mass_CO2 / molar_mass_CH4
-atmospheric_conc_per_mass_N2O = (1 * u.ppm) / (7.8 * u.gigatonne_N2O) * molar_mass_CO2 / molar_mass_N2O
-atmospheric_conc_per_mass_HFC = (1 * u.ppb) / (18.0 * u.megatonne_HFC) # HFC-134a (most common HFC)
-atmospheric_conc_per_mass_PFC = (1 * u.ppb) / (15.6 * u.megatonne_PFC) # CF4 (most common PFC)
-atmospheric_conc_per_mass_SF6 = (1 * u.ppb) / (25.9 * u.megatonne_SF6)
-atmospheric_conc_per_mass_NF3 = (1 * u.ppb) / (12.6 * u.megatonne_NF3)
-
-
-deltaF_coef_N2O = 0.12 * u.watt / (u.m * u.m) * surface_area_of_earth
-deltaF_coef_HFC = 0.16 * u.watt / (u.m * u.m) * surface_area_of_earth
-deltaF_coef_PFC = 0.08 * u.watt / (u.m * u.m) * surface_area_of_earth
-deltaF_coef_SF6 = 0.57 * u.watt / (u.m * u.m) * surface_area_of_earth
-deltaF_coef_NF3 = 0.21 * u.watt / (u.m * u.m) * surface_area_of_earth
-
-
-# TODO: use values in .ghgvalues
-CO2_GWP_100 = 1.0 * u.kg_CO2e / u.kg_CO2
-CH4_GWP_100 = 28.0 * u.kg_CO2e / u.kg_CH4
-N2O_GWP_100 = 265.0 * u.kg_CO2e / u.kg_N2O
-HFC_GWP_100 = 1_430 * u.kg_CO2e / u.kg_HFC
-PFC_GWP_100 = 6_630 * u.kg_CO2e / u.kg_PFC
-SF6_GWP_100 = 23_500 * u.kg_CO2e / u.kg_SF6
-NF3_GWP_100 = 16_100 * u.kg_CO2e / u.kg_NF3
-
-
-
-class AtmosphericChemistry(BaseScenarioProject):
-    """Combine GHG emissions into a CO2e estimate using GWP-100 emission factors,
-    and also simulate a simple radiative forcing and planetary heating model.
-    """
-    methane_decay_timescale:float = 10.0
-
-    may_register_emissions:bool = False
-    requires_emissions_registration_closed:bool = True
-
-    stepsize:object
-    decay_N2O:object
-    decay_HFC:object
-    decay_PFC:object
-    decay_SF6:object
-    decay_NF3:object
-
-    def __init__(self, stepsize=1.0 * u.years):
-        super().__init__(
-            stepsize=stepsize,
-            decay_N2O=(1 - stepsize / (114 * u.years)),
-            decay_HFC=(1 - stepsize / (14 * u.years)),
-            decay_PFC=1.0,
-            decay_SF6=1.0,
-            decay_NF3=1.0,
-            )
-
-    def on_add_project(self, state):
-        # TODO: use new ObjTensor support to vectorize this code
-
-        for catpath, contributors in state.sectoral_emissions_contributors.items():
-            # TODO: why not loop over these keys?
-            for sts_key in contributors.get(GHG.CO2, []):
-                state.declare_read_current_sts(self, sts_key)
-            for sts_key in contributors.get(GHG.CH4, []):
-                state.declare_read_current_sts(self, sts_key)
-            for sts_key in contributors.get(GHG.N2O, []):
-                state.declare_read_current_sts(self, sts_key)
-            for sts_key in contributors.get(GHG.HFCs, []):
-                state.declare_read_current_sts(self, sts_key)
-            for sts_key in contributors.get(GHG.PFCs, []):
-                state.declare_read_current_sts(self, sts_key)
-            for sts_key in contributors.get(GHG.SF6, []):
-                state.declare_read_current_sts(self, sts_key)
-            for sts_key in contributors.get(GHG.NF3, []):
-                state.declare_read_current_sts(self, sts_key)
-
-        with state.defining(self) as ctx:
-            for catpath, contributors in state.sectoral_emissions_contributors.items():
-                any_CO2e_contributors = False
-                if contributors.get(GHG.CO2, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_CO2_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_CO2, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if contributors.get(GHG.CH4, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_CH4_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_CH4, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if contributors.get(GHG.N2O, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_N2O_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_N2O, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if contributors.get(GHG.HFCs, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_HFC_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_HFC, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if contributors.get(GHG.PFCs, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_PFC_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_PFC, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if contributors.get(GHG.SF6, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_SF6_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_SF6, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if contributors.get(GHG.NF3, []):
-                    setattr(ctx, f'Predicted_Annual_Emitted_NF3_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_NF3, t_unit=u.year))
-                    any_CO2e_contributors = True
-                if any_CO2e_contributors:
-                    setattr(ctx, f'Predicted_Annual_Emitted_CO2e_mass_{catpath}',
-                            SparseTimeSeries(unit=u.kt_CO2e, t_unit=u.year))
-
-            ctx.Predicted_Annual_Emitted_CO2_mass = SparseTimeSeries(
-                unit=u.kt_CO2, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_CH4_mass = SparseTimeSeries(
-                unit=u.kt_CH4, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_N2O_mass = SparseTimeSeries(
-                unit=u.kt_N2O, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_HFC_mass = SparseTimeSeries(
-                unit=u.kt_HFC, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_PFC_mass = SparseTimeSeries(
-                unit=u.kt_PFC, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_SF6_mass = SparseTimeSeries(
-                unit=u.kt_SF6, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_NF3_mass = SparseTimeSeries(
-                unit=u.kt_NF3, interpolation='no_interpolation', t_unit=u.year)
-            ctx.Predicted_Annual_Emitted_CO2e_mass = SparseTimeSeries(
-                unit=u.kt_CO2e, interpolation='no_interpolation', t_unit=u.year)
-
-            # XXX : what year do these numbers represent? How can this be a default value
-            # when the simulated years are a parameter of the state?
-            ctx.Atmospheric_CO2_conc = SparseTimeSeries(unit=u.ppm, default_value=400.0 * u.ppm, t_unit=u.year)
-            ctx.Atmospheric_CH4_conc = SparseTimeSeries(unit=u.ppb, default_value=1775.0 * u.ppb, t_unit=u.year)
-            ctx.Atmospheric_N2O_conc = SparseTimeSeries(unit=u.ppb, default_value=336.0 * u.ppb, t_unit=u.year)
-
-            # Gemini says these data are from NOAA and are accurate for January 2026
-            ctx.Atmospheric_HFC_conc = SparseTimeSeries(unit=u.ppb, default_value=0.1345 * u.ppb, t_unit=u.year)
-            ctx.Atmospheric_PFC_conc = SparseTimeSeries(unit=u.ppb, default_value=0.0902 * u.ppb, t_unit=u.year)
-            ctx.Atmospheric_SF6_conc = SparseTimeSeries(unit=u.ppb, default_value=0.0124 * u.ppb, t_unit=u.year)
-            ctx.Atmospheric_NF3_conc = SparseTimeSeries(unit=u.ppb, default_value=0.0036 * u.ppb, t_unit=u.year)
-
-            ctx.DeltaF_CO2 = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_CH4 = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_N2O = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_HFC = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_PFC = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_SF6 = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_NF3 = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_forcing = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-            ctx.DeltaF_feedback = SparseTimeSeries(unit=u.petawatt, t_unit=u.year)
-
-            # Heat Energy forcing is the heat equivalent to net annual cashflow, an annual integral
-            ctx.Annual_Heat_Energy_forcing = SparseTimeSeries(default_value=0 * u.exajoule, t_unit=u.year)
-            ctx.Cumulative_Heat_Energy_forcing = SparseTimeSeries(default_value=0 * u.exajoule, t_unit=u.year)
-            ctx.Heat_Energy_imbalance = SparseTimeSeries(unit=u.exajoule, t_unit=u.year)
-            ctx.Cumulative_Heat_Energy = SparseTimeSeries(default_value=0.0 * u.exajoule, t_unit=u.year)
-            ctx.Ocean_Temperature_Anomaly = SparseTimeSeries(default_value=1.3 * u.kelvin, t_unit=u.year)
-
-        return int(state.t_now.to(u.years).magnitude + 1) * u.years
-
-    def step(self, state, current):
-
-        # add up annual emissions from registry
-        annual_CO2_mass = 0 * u.kt_CO2
-        annual_CH4_mass = 0 * u.kt_CH4
-        annual_N2O_mass = 0 * u.kt_N2O
-        annual_HFC_mass = 0 * u.kt_HFC
-        annual_PFC_mass = 0 * u.kt_PFC
-        annual_SF6_mass = 0 * u.kt_SF6
-        annual_NF3_mass = 0 * u.kt_NF3
-
-        for catpath, contributors in state.sectoral_emissions_contributors.items():
-            catpath_CO2e_mass = 0 * u.kg_CO2e
-            any_CO2e_contributors = False
-
-            catpath_CO2_contributors = contributors.get(GHG.CO2, [])
-            if catpath_CO2_contributors:
-                catpath_CO2_mass = sum(getattr(current, sts_key) for sts_key in catpath_CO2_contributors)
-                try:
-                    setattr(current, f'Predicted_Annual_Emitted_CO2_mass_{catpath}', catpath_CO2_mass)
-                except:
-                    print(catpath_CO2_contributors)
-                    raise
-                catpath_CO2e_mass += catpath_CO2_mass * CO2_GWP_100
-                annual_CO2_mass += catpath_CO2_mass
-                any_CO2e_contributors = True
-
-            catpath_CH4_contributors = contributors.get(GHG.CH4, [])
-            if catpath_CH4_contributors:
-                catpath_CH4_mass = sum(getattr(current, sts_key) for sts_key in catpath_CH4_contributors)
-                try:
-                    setattr(current, f'Predicted_Annual_Emitted_CH4_mass_{catpath}', catpath_CH4_mass)
-                except:
-                    print(catpath_CH4_contributors)
-                    raise
-                catpath_CO2e_mass += catpath_CH4_mass * CH4_GWP_100
-                annual_CH4_mass += catpath_CH4_mass
-                any_CO2e_contributors = True
-
-            catpath_N2O_contributors = contributors.get(GHG.N2O, [])
-            if catpath_N2O_contributors:
-                catpath_N2O_mass = sum(getattr(current, sts_key) for sts_key in catpath_N2O_contributors)
-                try:
-                    setattr(current, f'Predicted_Annual_Emitted_N2O_mass_{catpath}', catpath_N2O_mass)
-                except:
-                    print(catpath_N2O_contributors)
-                    raise
-                catpath_CO2e_mass += catpath_N2O_mass * N2O_GWP_100
-                annual_N2O_mass += catpath_N2O_mass
-                any_CO2e_contributors = True
-
-            catpath_HFC_contributors = contributors.get(GHG.HFCs, [])
-            if catpath_HFC_contributors:
-                catpath_HFC_mass = sum(getattr(current, sts_key) for sts_key in catpath_HFC_contributors)
-                setattr(current, f'Predicted_Annual_Emitted_HFC_mass_{catpath}', catpath_HFC_mass)
-                catpath_CO2e_mass += catpath_HFC_mass * HFC_GWP_100
-                annual_HFC_mass += catpath_HFC_mass
-                any_CO2e_contributors = True
-
-            catpath_PFC_contributors = contributors.get(GHG.PFCs, [])
-            if catpath_PFC_contributors:
-                catpath_PFC_mass = sum(getattr(current, sts_key) for sts_key in catpath_PFC_contributors)
-                setattr(current, f'Predicted_Annual_Emitted_PFC_mass_{catpath}', catpath_PFC_mass)
-                catpath_CO2e_mass += catpath_PFC_mass * PFC_GWP_100
-                annual_PFC_mass += catpath_PFC_mass
-                any_CO2e_contributors = True
-
-            catpath_SF6_contributors = contributors.get(GHG.SF6, [])
-            if catpath_SF6_contributors:
-                catpath_SF6_mass = sum(getattr(current, sts_key) for sts_key in catpath_SF6_contributors)
-                setattr(current, f'Predicted_Annual_Emitted_SF6_mass_{catpath}', catpath_SF6_mass)
-                catpath_CO2e_mass += catpath_SF6_mass * SF6_GWP_100
-                annual_SF6_mass += catpath_SF6_mass
-                any_CO2e_contributors = True
-
-            catpath_NF3_contributors = contributors.get(GHG.NF3, [])
-            if catpath_NF3_contributors:
-                catpath_NF3_mass = sum(getattr(current, sts_key) for sts_key in catpath_NF3_contributors)
-                setattr(current, f'Predicted_Annual_Emitted_NF3_mass_{catpath}', catpath_NF3_mass)
-                catpath_CO2e_mass += catpath_NF3_mass * NF3_GWP_100
-                annual_NF3_mass += catpath_NF3_mass
-                any_CO2e_contributors = True
-
-            if any_CO2e_contributors:
-                setattr(current, f'Predicted_Annual_Emitted_CO2e_mass_{catpath}', catpath_CO2e_mass)
-
-
-        current.Predicted_Annual_Emitted_CO2_mass = annual_CO2_mass
-        current.Predicted_Annual_Emitted_CH4_mass = annual_CH4_mass
-        current.Predicted_Annual_Emitted_N2O_mass = annual_N2O_mass
-        current.Predicted_Annual_Emitted_HFC_mass = annual_HFC_mass
-        current.Predicted_Annual_Emitted_PFC_mass = annual_PFC_mass
-        current.Predicted_Annual_Emitted_SF6_mass = annual_SF6_mass
-        current.Predicted_Annual_Emitted_NF3_mass = annual_NF3_mass
-        current.Predicted_Annual_Emitted_CO2e_mass = (
-            CO2_GWP_100 * annual_CO2_mass
-            + CH4_GWP_100 * annual_CH4_mass
-            + N2O_GWP_100 * annual_N2O_mass
-            + HFC_GWP_100 * annual_HFC_mass
-            + PFC_GWP_100 * annual_PFC_mass
-            + SF6_GWP_100 * annual_SF6_mass
-            + NF3_GWP_100 * annual_NF3_mass
-        )
-
-        fraction_of_emitted_CO2_that_becomes_atmospheric = .45
-
-        # apply an atmospheric climate model
-        annual_CO2_mass_atmospheric = (
-            annual_CO2_mass
-            * fraction_of_emitted_CO2_that_becomes_atmospheric)
-        annual_CH4_mass_atmospheric = annual_CH4_mass * 1.0 # no such discounting of CH4
-
-        annual_emitted_CO2_in_atmosphere_as_concentration = (
-            annual_CO2_mass_atmospheric
-            * atmospheric_conc_per_mass_CO2)
-
-        annual_emitted_CH4_in_atmosphere_as_concentration = (
-            annual_CH4_mass_atmospheric
-            / (2.78 * u.megatonne_CH4 / u.ppb)
-        ).to(u.ppb)
-
-        # TODO this should be multiplied by stepsize, not 1 year implicitly,
-        #      and this process should be tested for robustness to step size
-        tau_ch4 = 12.0 # years
-        annual_ch4_to_co2_decay = (
-            state.latest.Atmospheric_CH4_conc
-            / tau_ch4)
-
-        current.Atmospheric_CH4_conc = (
-            state.latest.Atmospheric_CH4_conc
-            + annual_emitted_CH4_in_atmosphere_as_concentration
-            + 180 * u.ppb # baseline from other sources
-            - annual_ch4_to_co2_decay)
-
-        # no decay is assumed for CO2
-        current.Atmospheric_CO2_conc = (
-            state.latest.Atmospheric_CO2_conc
-            + annual_emitted_CO2_in_atmosphere_as_concentration
-            + 2 * u.ppm # baseline from other sources
-            + (annual_ch4_to_co2_decay
-               * fraction_of_emitted_CO2_that_becomes_atmospheric)
-        )
-
-        reference_CO2_conc = 280.0 * u.ppm
-        current.DeltaF_CO2 = (
-            5.35 * u.watt / (u.m * u.m)
-            * surface_area_of_earth
-            * np.log(current.Atmospheric_CO2_conc.to(u.ppm).magnitude
-                     / reference_CO2_conc.to(u.ppm).magnitude))
-
-        reference_CH4_conc = 722.0 * u.ppb
-        current.DeltaF_CH4 = (
-            0.036 * u.watt / (u.m * u.m)
-            * surface_area_of_earth
-            * (np.sqrt(current.Atmospheric_CH4_conc.to(u.ppb).magnitude)
-               - np.sqrt(reference_CH4_conc.to(u.ppb).magnitude)))
-
-        self.step_N2O(state, current, annual_N2O_mass)
-        self.step_HFC(state, current, annual_HFC_mass)
-        self.step_PFC(state, current, annual_PFC_mass)
-        self.step_SF6(state, current, annual_SF6_mass)
-        self.step_NF3(state, current, annual_NF3_mass)
-
-        current.DeltaF_forcing = (
-            current.DeltaF_CO2
-            + current.DeltaF_CH4
-            + current.DeltaF_N2O
-            + current.DeltaF_HFC
-            + current.DeltaF_PFC
-            + current.DeltaF_SF6
-            + current.DeltaF_NF3
-        )
-
-        current.DeltaF_feedback = (
-            -1.3 * u.watt / (u.m * u.m) / u.kelvin
-            * surface_area_of_earth
-            * state.latest.Ocean_Temperature_Anomaly)
-
-        current.Annual_Heat_Energy_forcing = (
-            self.stepsize # integrate over duration of stepsize aka 1 year
-            * current.DeltaF_forcing)
-
-        current.Cumulative_Heat_Energy_forcing = (
-            state.latest.Cumulative_Heat_Energy_forcing
-            + self.stepsize # integrate over duration of stepsize aka 1 year
-            * current.DeltaF_forcing)
-
-        current.Heat_Energy_imbalance = (
-            self.stepsize # integrate over duration of stepsize aka 1 year
-            * (current.DeltaF_forcing + current.DeltaF_feedback))
-
-        specific_heat_of_top_200m_of_ocean = 151200.0 * u.exajoule / u.kelvin * 2
-        current.Ocean_Temperature_Anomaly = (
-            state.latest.Ocean_Temperature_Anomaly
-            + (current.Heat_Energy_imbalance
-               / (specific_heat_of_top_200m_of_ocean)))
-
-        current.Cumulative_Heat_Energy = (
-            state.latest.Cumulative_Heat_Energy
-            + current.Heat_Energy_imbalance)
-
-        return int(state.t_now.to(u.years).magnitude + 1) * u.years
-
-    def step_N2O(self, state, current, annual_N2O_mass):
-        conc = state.latest.Atmospheric_N2O_conc
-
-        conc += atmospheric_conc_per_mass_N2O * annual_N2O_mass
-        conc *= self.decay_N2O
-        current.Atmospheric_N2O_conc = conc
-
-        reference_N2O_conc = 270.0 * u.ppb
-
-        current.DeltaF_N2O = (
-            deltaF_coef_N2O
-            * (np.sqrt(conc.to(u.ppb).magnitude)
-               - np.sqrt(reference_N2O_conc.to(u.ppb).magnitude)))
-
-    def step_HFC(self, state, current, annual_HFC_mass):
-        conc = state.latest.Atmospheric_HFC_conc
-
-        conc += atmospheric_conc_per_mass_HFC * annual_HFC_mass
-        conc *= self.decay_HFC
-        current.Atmospheric_HFC_conc = conc
-
-        current.DeltaF_HFC = deltaF_coef_HFC * conc.to(u.ppb).magnitude
-
-    def step_PFC(self, state, current, annual_PFC_mass):
-        conc = state.latest.Atmospheric_PFC_conc
-
-        conc += atmospheric_conc_per_mass_PFC * annual_PFC_mass
-        conc *= self.decay_PFC
-        current.Atmospheric_PFC_conc = conc
-
-        current.DeltaF_PFC = deltaF_coef_PFC * conc.to(u.ppb).magnitude
-
-    def step_SF6(self, state, current, annual_SF6_mass):
-        conc = state.latest.Atmospheric_SF6_conc
-
-        conc += atmospheric_conc_per_mass_SF6 * annual_SF6_mass
-        conc *= self.decay_SF6
-        current.Atmospheric_SF6_conc = conc
-
-        current.DeltaF_SF6 = deltaF_coef_SF6 * conc.to(u.ppb).magnitude
-
-    def step_NF3(self, state, current, annual_NF3_mass):
-        conc = state.latest.Atmospheric_NF3_conc
-
-        conc += atmospheric_conc_per_mass_NF3 * annual_NF3_mass
-        conc *= self.decay_NF3
-        current.Atmospheric_NF3_conc = conc
-
-        current.DeltaF_NF3 = deltaF_coef_NF3 * conc.to(u.ppb).magnitude
-
 class GeometricHumanPopulationForecast(BaseScenarioProject):
     rate:float = 1.014
     stepsize:object = 1.0 * u.years
@@ -1055,270 +624,6 @@ class GeometricHumanPopulationForecast(BaseScenarioProject):
         current.human_population *= self.rate
         return state.t_now + self.stepsize
 
-
-class ProjectComparison(object):
-    def __init__(self, state_A, state_B, present, project):
-        self.state_A = state_A # state with project
-        self.state_B = state_B # baseline state
-        self.present = present
-        self.project = project
-
-    def _years(self):
-        t_start = min(self.state_A.t_start, self.state_B.t_start)
-        t_stop = max(self.state_A.t_now, self.state_B.t_now)
-        start_year = int(t_start.to('years').magnitude)
-        stop_year = int(t_stop.to('years').magnitude) + 1
-        years = np.arange(start_year, stop_year) * u.years
-        return years
-
-    def years_as_list(self):
-        return [int(year.to('years').magnitude) for year in self._years()]
-
-    @property
-    def _present_year_int(self):
-        return int(self.present.to('years').magnitude)
-
-    def _net_present_envelope(self, years, base_rate):
-        present_year_int = self._present_year_int
-        envelope = [0] * len(years)
-        for ii, year in enumerate(years):
-            year_int = int(year.to('years').magnitude)
-            if year_int >= present_year_int:
-                envelope[ii] = base_rate ** (year_int - present_year_int)
-        return np.asarray(envelope)
-
-    def net_present_discounted_sum(self, base_rate, key):
-        years = self._years()
-        vals_A = self.state_A.sts[key].query(years)
-        vals_B = self.state_B.sts[key].query(years)
-        diff = vals_A - vals_B
-        envelope = self._net_present_envelope(years, base_rate)
-        return np.cumsum(diff.magnitude * envelope)[-1] * diff.u
-
-    def net_present_CO2e(self, base_rate):
-        return self.net_present_discounted_sum(
-            base_rate,
-            key='Predicted_Annual_Emitted_CO2e_mass')
-
-    def net_present_heat(self, base_rate):
-        return self.net_present_discounted_sum(
-            base_rate,
-            key='Annual_Heat_Energy_forcing')
-
-    def net_present_value(self, base_rate):
-        if self.project.after_tax_cashflow_name in self.state_B.sts:
-            raise NotImplementedError()
-        years = self._years()
-        envelope = self._net_present_envelope(years, base_rate)
-        cashflow = self.state_A.sts[self.project.after_tax_cashflow_name].query(years)
-        return np.cumsum(cashflow.magnitude * envelope)[-1] * cashflow.u
-
-    def cost_per_ton_CO2e(self, base_rate):
-        npv = self.net_present_value(base_rate=base_rate)
-        npc = self.net_present_CO2e(base_rate=base_rate)
-        if npc >= 0:
-            return float('nan') * u.CAD / u.tonne_CO2e
-        return (npv / npc).to(u.CAD / u.tonne_CO2e)
-
-    def echart_series_Mt(self, A_or_B, catpath, stack=None, name=None):
-        years = self._years()
-        if A_or_B == "A":
-            state = self.state_A
-        elif A_or_B == "B":
-            state = self.state_B
-        else:
-            raise NotImplementedError(A_or_B)
-        predictions = state.sts[f'Predicted_Annual_Emitted_CO2e_mass_{catpath}'].query(
-            years)
-        data = [{'value': float(datum.to(u.megatonne_CO2e).magnitude),
-                 'url': f'/ipcc-sectors/{catpath}'.replace(' ', '_')}
-                for datum in predictions]
-        rval = dict(
-            name=name or catpath,
-            type='line',
-            #areaStyle={},
-            #emphasis={'focus': 'series'},
-            data=data)
-        if stack:
-            rval['stack'] = stack
-        return rval
-
-
-class ProjectEvaluation(object):
-    def __init__(self, projects, common_projects, alt_project=None, present=None):
-        self.projects = projects # dict
-        self.common_projects = common_projects
-        self.present = (
-            time.time() * u.seconds + 1970 * u.years
-            if present is None else present)
-
-        self.comparisons = {}
-        self.states = {}
-        default_state = None
-        for eval_name, prj in projects.items():
-            if isinstance(prj, (list, tuple)):
-                raise NotImplementedError()
-            else:
-                state_A = State(name=f'StateA_{eval_name}')
-                state_A.add_project(prj)
-                state_A.add_projects(common_projects)
-                if default_state is None:
-                    default_state = State(name=f'Baseline')
-                    default_state.add_projects(common_projects)
-                self.comparisons[eval_name] = ProjectComparison(
-                    state_A=state_A,
-                    state_B=default_state,
-                    present=self.present,
-                    project=prj)
-                self.states[state_A.name] = state_A
-                self.states[default_state.name] = default_state
-
-    def run_until(self, t_stop):
-        for state in self.states.values():
-            state.run_until(t_stop)
-
-    def all_sts_names(self):
-        rval = set()
-        for state in self.states.values():
-            rval.update(state.sts.keys())
-        return rval
-
-    def plot(self, t_unit='years', **kwargs):
-
-        sorted_sts_names = list(sorted(self.all_sts_names()))
-
-        if len(sorted_sts_names) <= 1:
-            fig = plt.figure()
-            rows = 1
-            cols = 1
-        elif len(sorted_sts_names) <= 6:
-            fig = plt.figure(figsize=[12, 8])
-            rows = 2
-            cols = 3
-        elif len(sorted_sts_names) <= 15:
-            rows = 5
-            cols = 3
-            fig = plt.figure(figsize=[12, (len(sorted_sts_names) // cols + 1) * 5.5])
-        elif len(sorted_sts_names) <= 28:
-            rows = 7
-            cols = 4
-            fig = plt.figure(figsize=[15, (len(sorted_sts_names) // cols + 1) * 5.5])
-        else:
-            raise NotImplementedError()
-        fig.set_layout_engine("constrained")
-
-        for ii, sts_name in enumerate(sorted_sts_names):
-            plt.subplot(rows, cols, ii + 1)
-            for state in self.states.values():
-                if sts_name in state.sts:
-                    state.sts[sts_name].plot(t_unit=t_unit, annotate=True, label=state.name)
-
-    def plot_nph_vs_npv(self, discount_rate, nph_unit='exajoule', npv_unit='MCAD'):
-        base_rate = (1 - discount_rate)
-        eval_names = []
-        nph1s = []
-        npv1s = []
-        for eval_name, cmp in self.comparisons.items():
-            nph1 = cmp.net_present_heat(base_rate=base_rate)
-            npv1 = cmp.net_present_value(base_rate=base_rate)
-            eval_names.append(eval_name)
-            nph1s.append(nph1.to(nph_unit).magnitude)
-            npv1s.append(npv1.to(npv_unit).magnitude)
-
-        plt.figure()
-        plt.title(f'Future-Discounted Project Comparison @ {discount_rate * 100:.1f}% ({base_rate ** 100:.2f} at 100 years)')
-        plt.scatter(npv1s, nph1s)
-        for ii, eval_name in enumerate(eval_names):
-            plt.annotate(eval_name, (npv1s[ii], nph1s[ii]))
-        plt.xlabel(f'Net Present Value ({npv_unit})')
-        plt.ylabel(f'Net Present Heat Forcing ({nph_unit})')
-
-    def iter_npv_nph_evalname(self, discount_rate):
-        base_rate = (1 - discount_rate)
-        for eval_name, cmp in self.comparisons.items():
-            nph = cmp.net_present_heat(base_rate=base_rate)
-            npv = cmp.net_present_value(base_rate=base_rate)
-            yield npv, nph, eval_name
-
-
-class IPCC_Forest_Land_Model(BaseScenarioProject):
-    stepsize:object = 1.0 * u.years
-
-    def on_add_project(self, state):
-        with state.requiring_current(self) as ctx:
-
-            ctx.Other_Forest_Land_CO2 = SparseTimeSeries(
-                default_value=40.0 * u.Mt_CO2)
-
-        state.register_emission('Forest_Land', GHG.CO2, 'Other_Forest_Land_CO2')
-
-
-class IPCC_Transport_RoadTransportation_LightDutyGasolineTrucks(BaseScenarioProject):
-    stepsize:object = 1.0 * u.years
-
-    def on_add_project(self, state):
-        with state.requiring_current(self) as ctx:
-            ctx.human_population = SparseTimeSeries(
-                times=[state.t_now],
-                values=[27_685_730 * u.people])
-            ctx.Government_LightDutyGasolineTrucks_ZEV_fraction = SparseTimeSeries(
-                default_value=0 * u.dimensionless)
-            ctx.Other_LightDutyGasolineTrucks_ZEV_fraction = SparseTimeSeries(
-                default_value=0 * u.dimensionless)
-
-        with state.defining(self) as ctx:
-            # https://www.canada.ca/en/treasury-board-secretariat/services/innovation/greening-government/government-canada-greenhouse-gas-emissions-inventory.html
-            # not exactly using ^^ but guessing based on that and Gemini estimation
-
-            # TODO: factor in CO2e footprint of manufacturing each type of vehicle
-
-            # TODO: factor in the emissions of electricity generation in each province
-            #       and the population of each province
-
-            ctx.Government_LightDutyGasolineTrucks_CO2 = SparseTimeSeries(
-                default_value=0 * u.Mt_CO2)
-            ctx.Other_LightDutyGasolineTrucks_CO2 = SparseTimeSeries(
-                default_value=0 * u.Mt_CO2)
-
-        state.register_emission('Transport/Road_Transportation/Light-Duty_Gasoline_Trucks', GHG.CO2, 'Other_LightDutyGasolineTrucks_CO2')
-        state.register_emission('Transport/Road_Transportation/Light-Duty_Gasoline_Trucks', GHG.CO2, 'Government_LightDutyGasolineTrucks_CO2')
-        return state.t_now + self.stepsize
-
-    def step(self, state, current):
-        coefficient = 1_200 * u.kg_CO2 / u.people
-        current.Government_LightDutyGasolineTrucks_CO2 = (
-            current.human_population * coefficient * .025
-            * (1 * u.dimensionless - current.Government_LightDutyGasolineTrucks_ZEV_fraction))
-        current.Other_LightDutyGasolineTrucks_CO2 = (
-            current.human_population * coefficient * .975
-            * (1 * u.dimensionless - current.Other_LightDutyGasolineTrucks_ZEV_fraction))
-        return state.t_now + self.stepsize
-
-
-class SubsidyAccounting(BaseScenarioProject):
-    """Tally up annual subsidy amounts required by barriers and strategies."""
-
-    def on_add_project(self, state):
-        for sts_key in state.subsidy_requirements:
-            state.declare_read_current_sts(self, sts_key)
-        with state.defining(self) as ctx:
-            ctx.AnnualSubsidyTotal = STS(
-                times=array.array('d', []),
-                values=array.array('d', [float('nan')]),
-                t_unit=u.year,
-                v_unit=u.CAD,
-                interpolation='no_interpolation')
-        return state.t_now.to('year').magnitude * u.year
-
-    def step(self, state, current):
-        zero = 0 * u.CAD
-        total = zero
-        for sts_key in state.subsidy_requirements:
-            subtotal = getattr(current, sts_key)
-            assert subtotal >= zero # sign convention and nan-check
-            total += subtotal
-        current.AnnualSubsidyTotal = total
-        return state.t_now + 1 * u.year
 
 from . import ipcc_canada
 from . import ghgvalues
