+++ planzero/sim.py
@@ -7,6 +7,7 @@ import numpy as np
 
 from .ureg import u
 from .base import State
+from .base import Other_NIR_Historical_Actuals
 
 from .html import (
     EChartTitle,
@@ -204,12 +205,6 @@ class SiteSimulation(BaseModel):
                 if de.identifier == identifier]
 
 
-from .base import (
-    AtmosphericChemistry,
-    SubsidyAccounting,
-    Other_NIR_Historical_Actuals,
-    )
-
 
 class Extrapolation(SiteSimulation):
     """Extend statistical trends in emissions contributions"""
@@ -249,9 +244,8 @@ class Scaling(SiteSimulation):
 
             # standard for vis
             Other_NIR_Historical_Actuals(),
-            #AtmosphericChemistry(),
-            #SubsidyAccounting(),
         ]
+
 
 @cache
 def simulation_result(simulation_name) -> SimulationResult:
